CHT-571. Remove usages of getTag() and getAttrString() which are being removed from the SDK

See merge request megachat/MEGAchat!1250

+++ src/base/promise.h
@@ -178,7 +178,7 @@ public:
     }
     inline int count() const
     {
-        return items.size();
+        return static_cast<int>(items.size());
     }
     inline void addListMoveItems(CallbackList& other)
     {
